Bring the NimbusEffect into sst-effects
This ports the nimbus effect from Surge from a DSP and
params perspective, but doesn't impement the dynamic param
names by engine type feature yet. That's a separate commit
in the future once I figure out how I want to do that.

+++ include/sst/effects/EffectCore.h
@@ -171,10 +171,7 @@ template <typename FXConfig> struct EffectTemplateBase : public FXConfig::BaseCl
         }
     }
 
-    inline float intValue(int idx) const
-    {
-        return FXConfig::intValueAt(asBase(), valueStorage, idx);
-    }
+    inline int intValue(int idx) const { return FXConfig::intValueAt(asBase(), valueStorage, idx); }
 
     inline float temposyncRatio(int idx) const
     {
